Refactor render page view into funtion
